Created BL PulseManager and BL Test utPulseManager

+++ LungTracking/LungTracking.BL.Models/Pulse.cs
@@ -9,7 +9,7 @@ namespace LungTracking.BL.Models
     public class Pulse
     {
         public Guid Id { get; set; }
-        public decimal PulseNumber { get; set; }
+        public int PulseNumber { get; set; }
         public bool BeginningEnd { get; set; }
         public DateTime TimeOfDay { get; set; }
         public Guid PatientId { get; set; }
